+++ src/container/modules/node-transformers/SimplifyingTransformersModule.ts
@@ -5,11 +5,11 @@ import { INodeTransformer } from '../../../interfaces/node-transformers/INodeTra
 
 import { NodeTransformer } from '../../../enums/node-transformers/NodeTransformer';
 
-import { IfStatementSimplifyTransformer } from '../../../node-transformers/minification-transformers/IfStatementSimplifyTransformer';
-import { VariableDeclarationsMergeTransformer } from '../../../node-transformers/minification-transformers/VariableDeclarationsMergeTransformer';
+import { IfStatementSimplifyTransformer } from '../../../node-transformers/simplifying-transformers/IfStatementSimplifyTransformer';
+import { VariableDeclarationsMergeTransformer } from '../../../node-transformers/simplifying-transformers/VariableDeclarationsMergeTransformer';
 
-export const minificationTransformersModule: interfaces.ContainerModule = new ContainerModule((bind: interfaces.Bind) => {
-    // minification transformers
+export const simplifyingTransformersModule: interfaces.ContainerModule = new ContainerModule((bind: interfaces.Bind) => {
+    // simplifying transformers
     bind<INodeTransformer>(ServiceIdentifiers.INodeTransformer)
         .to(IfStatementSimplifyTransformer)
         .whenTargetNamed(NodeTransformer.IfStatementSimplifyTransformer);