+++ src/enums/node-transformers/NodeTransformationStage.ts
@@ -6,6 +6,6 @@ export enum NodeTransformationStage {
     RenameProperties = 'RenameProperties',
     Converting = 'Converting',
     Obfuscating = 'Obfuscating',
-    Minification = 'Minification',
+    Simplifying = 'Simplifying',
     Finalizing = 'Finalizing'
 }