+++ test/functional-tests/javascript-obfuscator/JavaScriptObfuscator.spec.ts
@@ -789,7 +789,7 @@ describe('JavaScriptObfuscator', () => {
                         deadCodeInjection: true,
                         deadCodeInjectionThreshold: 1,
                         disableConsoleOutput: false,
-                        minify: true,
+                        simplify: true,
                         renameProperties: true,
                         rotateStringArray: true,
                         stringArray: true,